+++ src/rules.js
@@ -1,5 +1,6 @@
 export default {
 	indent: 2,
+	blankLines: '\n',
 	eof: '\n',
 	eol: '\n'
 };