+++ test/test-plugin.js
@@ -1,56 +1,27 @@
 import {readFile} from 'fs';
 import test from 'ava';
-import posthtml from 'posthtml';
 import isPromise from 'is-promise';
+import {processing, read} from './utils';
 import beautify from '../src';
 
-function processing(html, plugins = [], options = {}) {
-	return posthtml(plugins)
-		.process(html, options);
-}
-
-function read(pathFile) {
-	return new Promise((resolve, reject) => {
-		readFile(pathFile, 'utf8', (err, data) => {
-			if (err) {
-				reject(err);
-			}
-			return resolve(data);
-		});
-	});
-}
-
-test('processing should return promise', t => {
-	t.true(isPromise(processing('')));
-});
-
-test('processing not should return promise with use option sync', t => {
-	t.false(isPromise(processing('<div></div>', [], {sync: true})));
-});
-
-test('processing should return object with use option sync', t => {
-	t.deepEqual(typeof processing('<div></div>', [], {sync: true}), 'object');
-});
-
 test('plugin beautify must be function', t => {
+	t.plan(2);
 	t.true(typeof beautify === 'function');
-});
-
-test('plugin beautify not should return promise with use option sync', t => {
-	t.false(isPromise(beautify({sync: true})()));
+	t.true(typeof beautify() === 'function');
 });
 
 test('plugin beautify should return Error with use option sync', t => {
-	t.true(beautify({sync: true})().name === 'Error');
+	const error = t.throws(() => {
+		beautify()();
+	}, TypeError);
+
+	t.is(error.message, 'tree is not Array');
 });
 
 test('plugin beautify should return the passed value with use option sync', t => {
-	t.deepEqual([], beautify({sync: true})([]));
-});
+	const tree = [];
 
-test('processing with plugin beautify should return equal html with use option sync', async t => {
-	const fixture = '<div></div>\n\n<div></div>';
-	t.deepEqual(fixture, (await processing(fixture, [beautify({sync: true, rules: {eof: false}})], {sync: true}).html));
+	t.deepEqual(tree, beautify()(tree));
 });
 
 test('processing with plugin beautify should not lost native api', async t => {
@@ -58,19 +29,19 @@ test('processing with plugin beautify should not lost native api', async t => {
 	const fixture = '<div></div>\n\n<div></div>';
 	t.true(
 		Object.prototype.hasOwnProperty.call(
-			await processing(fixture, [beautify({sync: true, rules: {eof: false}})], {sync: true}).tree,
+			(await processing(fixture, [beautify({rules: {eof: false}})])).tree,
 			'walk'
 		)
 	);
 	t.true(
 		Object.prototype.hasOwnProperty.call(
-			await processing(fixture, [beautify({sync: true, rules: {eof: false}})], {sync: true}).tree,
+			(await processing(fixture, [beautify({rules: {eof: false}})])).tree,
 			'match'
 		)
 	);
 	t.true(
 		Object.prototype.hasOwnProperty.call(
-			await processing(fixture, [beautify({sync: true, rules: {eof: false}})], {sync: true}).tree,
+			(await processing(fixture, [beautify({rules: {eof: false}})])).tree,
 			'processor'
 		)
 	);
@@ -106,8 +77,8 @@ test('processing with plugin beautify should return boolean attribute', async t 
 });
 
 test('processing with plugin beautify should return trim attribute', async t => {
-	const fixture = '<span class=" image  "  rel="  images "></span>';
-	const expected = '<span class="image" rel="images"></span>';
+	const fixture = '<span class=" image  response"  rel="  images "></span>';
+	const expected = '<span class="image response" rel="images"></span>';
 	t.deepEqual(expected, (await processing(fixture, [beautify({rules: {eof: false}})])).html);
 });
 
@@ -127,6 +98,13 @@ test('processing with plugin beautify should return with indent', async t => {
 	t.deepEqual(
 		(await read('test/expected/output-indent.html')),
 		(await processing(await read('test/fixtures/input-indent.html'), [beautify()])).html
+	);
+});
+
+test('processing with plugin beautify should return with without blankLines', async t => {
+	t.deepEqual(
+		(await read('test/expected/output-blank-lines.html')),
+		(await processing(await read('test/fixtures/input-blank-lines.html'), [beautify({rules: {blankLines: false}})])).html
 	);
 });